server/requirements.txt
future==0.18.2
imutils==0.5.4
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pynput==1.7.3
pyobjc-core==7.1
pyobjc-framework-Cocoa==7.1
pyobjc-framework-Quartz==7.1
six==1.15.0
torch==1.6.0
torchvision==0.7.0
typing-extensions==3.7.4.3
websockets==8.1
